update argument size to allow addresses to be passed
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/ec/trunk@35758 72102866-910b-0410-8b05-ffd578937521

+++ EcCommand.m
@@ -113,7 +113,7 @@ static NSString*	cmdWord(NSArray* a, unsigned int pos)
 - (void) cmdPing: (id <CmdPing>)from
 	sequence: (unsigned)num
 	   extra: (NSData*)data;
-- (void) cmdQuit: (int)sig;
+- (void) cmdQuit: (NSInteger)sig;
 - (void) command: (NSData*)dat
 	      to: (NSString*)t
 	    from: (NSString*)f;
@@ -507,7 +507,7 @@ static NSString*	cmdWord(NSArray* a, unsigned int pos)
   [from cmdGnip: self sequence: num extra: nil];
 }
 
-- (void) cmdQuit: (int)sig
+- (void) cmdQuit: (NSInteger)sig
 {
   if (sig == tStatus && control != nil)
     {
