+++ EcControl.m
@@ -336,7 +336,7 @@ static NSString*	cmdWord(NSArray* a, unsigned int pos)
 - (void) cmdPing: (id <CmdPing>)from
 	sequence: (unsigned)num
 	   extra: (NSData*)data;
-- (void) cmdQuit: (int)status;
+- (void) cmdQuit: (NSInteger)status;
 - (void) command: (NSData*)dat
 	    from: (NSString*)f;
 - (BOOL) connection: (NSConnection*)ancestor
@@ -526,7 +526,7 @@ static NSString*	cmdWord(NSArray* a, unsigned int pos)
   [from cmdGnip: self sequence: num extra: nil];
 }
 
-- (void) cmdQuit: (int)status
+- (void) cmdQuit: (NSInteger)status
 {
   [sink shutdown];
   exit (status);
